requirements.txt
docopt==0.6.2
pep8==1.7.0
terminaltables==3.1.0
GitPython==2.1.1
